requirements.txt
yfinance>=1.2.0
pandas==2.1.4
psycopg2-binary==2.9.9
sqlalchemy==1.4.51
apache-airflow-providers-postgres==5.10.0
matplotlib==3.8.2
